my-pesa-backend/requirements.txt
asgiref==3.6.0
boto3==1.26.67
botocore==1.29.67
Django==4.1.6
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
Pillow==9.4.0
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
